recipe_backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
Events==0.5
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
opensearch-py==2.7.1
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==74.1.2
six==1.16.0
tldextract==5.1.2
urllib3==2.2.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3